+++ src/routes/cms.routes.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const { submitCmsIntake } = require('../controllers/cms.controller');
-
-const router = express.Router();
-
-router.post('/submit', submitCmsIntake);
-
-module.exports = router;